Fix "ghost" barrier and focus effects.
A barrier or focus effect may be displayed on a unit that was
barriered or focused at the beginning of the turn, but had been
dispelled by an attack.  If that attack involved luck and an undo
took place, then it would display the barrier and/or focus even
though the attack had dispelled them.  This did not affect gameplay
since the unit is not actually barriered and/or focused, but can be
visually confusing.