Improve idle connection stability
It is not uncommon for a single message to get lost in a websocket
connection.  This is why sync messages are sent every 5 seconds to
ensure lost messages are resent in a timely manner.  It is also
possible for the connection to be lost without knowing it.  This is
why a connection is reset if nothing is received after a while.
The connection timeout WAS 10 seconds.  This means a single lost
sync message can result in a connection reset.  So, the connection
timeout was raised to 15 seconds.  Now two sync messages would have
to be lost to cause a connection reset.  This way a connection is
not assumed to be lost if it was merely a message that was lost.