Atomically create and revoke identity tokens
There is a case where a profile can be saved and a token created or
revoked in parallel.  Since creating and revoking identity tokens
wasn't atomic, it was reading the player data before it was updated
by saving the profile and then written after the profile was saved,
erasing the assigned nextToken.  This would cause a Revoked token
error when the client would try to use/activate their new token.