Fetch games on interval and repaint game lists
This changes the main render function so that it will clear the game lists (now
given another class of `"gameList"` so they can be easily selected) before
rendering them.

Also, after the initial fetch and render, an interval will be setup that will
fetch and render the games every 30 seconds.