Registration throttle reduced to 30 seconds
Stupid iOS Safari treats a PWA and the browser as two separate
containers.  They do not share local storage and newer iOS versions
don't share anything else.  So, it is highly probable that an
account would be registered twice in short order on such devices.
This resulted in hitting the registration throttle when it was set
to 60 seconds.  Reducing it should help.  But ultimately I need to
find a way to avoid confusion on iPhones and guide the user to have
only one account.