Rework account recovery and testing strategy
An identity token can be generated for a player's account and used
to create a new device.  This is a manual way to restore a player's
account if they lose it.  Or, it can be used to gain access to a
player's account to test bugs.  In the latter case and if there is
an existing unused identity token, be sure to restore it after use.