Fix game stuck condition
The game would get stuck if you leave the game tab during your turn
then return.  This is because the RemoteTransport emits a 'change'
event when reconnecting to the game.  If no change had occurred,
then the Game.play() leaves the board in a locked state.  So, I
changed the "GameStateCursor.atEnd" method to only return true if
the game state is also in sync.  This made it simple to ignore
calls to Game.play() if there is nothing more to play.