Don't merge arrays.  Replace them.
This fixes a bug where somebody resumes a game and the only change
in state is additional actions and end-of-turn.  The new state is
merged with the old state, resulting in merging the current non-
empty array with an empty array.  The remote transport would then
have the wrong actions in its state, which can cause the client to
crash.