Simplified cursor and play management
I noticed that sometimes the client would request turn actions for
the same turn 3 times in rapid succession.  Not being efficient, it
was more concerning that it was unexpected.  The solution was to
combine isOutOfSync(), sync(), and setNextAction() into a single
method that only fetched the data once and did the right thing.

Also, the play stack didn't end up being a stack, so it was changed
to use a single promise that encapsulates the entire play process.
This is also a better guarantee that there won't be parallel play
activity.