Don't send queued messages until session opens
When a connection is interrupted, session data is not cleared
until it fails to resume.  Since session data is present, messages
may be sent between connecting and resuming a session.  This was
observed on the game list page where it would search for player
games every 5 seconds.  Sometimes the query would be sent before
resuming the session and would result in an error.  So, the _send()
method should check if the session is open rather than checking for
the presence of session data.