Indicate 'No Luck' games in active list
Fixed a bug that only affects replay mode after your game ends and
a surrender took place and it was the surrendering team's turn and
they had taken an action before surrendering.  The game state would
never progress to the empty last turn.  So, when resuming replay
from the end of the game, it would try to select a unit that had
already died.

The fix involved moving client state to a new turn immediately upon
receiving an endTurn action.  It does not wait for a 'startTurn'
event that will never come when the game ends.