Add unit info card to set setup page
DSM power modifiers are now calculated during set setup so they may
be observed in the unit info card.

DSM power modifiers are now saved against the set instead of being
initialized by game state at start of game.

mHealth now has a minimum cap, which is how it worked originally.
Instead, the total damage (or healing) is captured in a separate
field so that it may be displayed while playing out an attack.