Replace existing game when creating same type
If you attempt to create a public game that is a duplicate of an
existing waiting game, it will redirect you to the existing game
as it did before.

But if you attempt to create a public game that has the same type
as an existing waiting game, but with different settings, then it
will cancel the existing game as it creates a new one.