Fixed undo bug
If the opponent makes a move, ends their turn, hits undo, then does
something different and ends their turn quick enough, you may fail
to see the change in moves.  A fix is now in place to ensure that
the actions seen this turn are still current despite the fact that
the next turn has started.  This is done by comparing the date of
the last action this turn with the date the next turn started. This
is now guaranteed to be the same if everything is in sync.