Fix fork game cheat in luck-based games
Also refactored/moved fork logic to Game model.

In luck-based games, it was possible to figure out if an attack
will get blocked or not.  All you had to do was fork the game and
see what would happen.  This is due to seeded RNG.  The fix is to
generate a new random seed for the practice game.