Do not auto-replay from beginning for ongoing games
When hitting the "play" button in replay mode and the cursor is
already at the end of playback, i.e. the "current" game cursor, it
would reset playback to the beginning of the game and play from
there.  However, this may not be the desired operation if the game
hasn't ended yet.  Rather, you may wish to wait until the game
continues.