Further optimize serialized objects.
This change was motivated by the fact that old clients could not detect
the latest version.  This is because the open session response was
serialized and old clients don't know how to normalize.  In fact, the
open session response does not require serialization.  So, now if an
object does not require serialization, it is ignored.  This means the
serialization properties 'type', 'transform', and 'data' all need to be
prefixed with a $ so that they aren't mistaken for data properties.