Fix phantom games
Games were appearing and immediately disappearing from the lobby if
their time limit expired and one of the players opens the game.
Opening a game changes its state since the last time the player
opened game is recorded.  But the game should not appear in the
lobby until the player makes their move, making it truly active.