Clear lobby game when clearing your lobby game
When a floor is full and you remove your avatar from the only
empty arena, the first event you receive is removing the game from
your "myGames" collection.  This clears your lobby game, but did
not clear the game from the lobby games list.  So "getLobbyGames"
thought the floor was truly full and would empty out the oldest
game from the floor.  But then we get another event removing your
game from the lobby style collection and it would restore the
previously cleared oldest game.  It isn't desireable for a game to
disappear and immediately reappear, so clearing the lobby game when
clearing your lobby game fixes this.